+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/PlaceObject2Tag.java
@@ -472,7 +472,12 @@ public class PlaceObject2Tag extends PlaceObjectTypeTag implements ASMSourceCont
     public void setPlaceFlagHasMatrix(boolean placeFlagHasMatrix) {
         this.placeFlagHasMatrix = placeFlagHasMatrix;
     }
-
+    
+    @Override
+    public void setPlaceFlagMove(boolean placeFlagMove) {
+        this.placeFlagMove = placeFlagMove;
+    }
+    
     @Override
     public boolean hasImage() {
         return false;